Numeric Precision Issue
• Fixes an issue where floating point session time trigger values were not being accounted for properly since they were being converted to long before being multiplied 1000 (to convert to milliseconds)

+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/OSDynamicTriggerController.java
@@ -42,7 +42,7 @@ class OSDynamicTriggerController {
             if (scheduledMessages.contains(trigger.triggerId))
                 return false;
 
-            long requiredTimeInterval = ((Number)trigger.value).longValue() * 1000;
+            long requiredTimeInterval = (long)(((Number)trigger.value).doubleValue() * 1000);
 
             long offset = 0;
